Refactor client connection handling and enhance Docker support
- Moved connection logic to a new `connection.go` file for better organization.
- Introduced Docker container tracking with `containerID` and `isDockerCommand` fields.
- Implemented methods for reading container IDs and killing Docker containers during disconnect.
- Enhanced logging for Docker operations and connection health checks.
- Added constants for transport types to improve code clarity.
