+++ internal/upstream/core/constants.go
@@ -1,0 +1,5 @@
+package core
+
+const (
+	transportStdio = "stdio"
+)
